backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.31.0
ipfshttpclient==0.8.0
web3==6.15.1
python-magic==0.4.27 